+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.8.25
 
 * Fixed false positive on transitionEnd events in Transition
+* Fixed issue with inaccessible Number inputs on iOS
 * Increased initial pointer down delay in Number 
 
 ### 0.8.24